+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+- If you have expanded config in your variant listing config and it is not active, it will now not have a look at those settings anymore. 
+
 ## [3.5.1] - 13-02-2025
 
 ### Fixed